GUI/requirements.txt
numpy>=1.19.0
scipy>=1.5.0
Pillow>=8.0.0
torch>=1.7.0
torchvision>=0.8.0
wxPython>=4.1.0
pydub>=0.25.0
pyaudio>=0.2.11


